# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BookingCreatorOut": {
        "properties": {
          "city": {
            "title": "City",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "specialty": {
            "title": "Specialty",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "city",
          "specialty"
        ],
        "title": "BookingCreatorOut",
        "type": "object"
      },
      "BookingEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BookingOut"
          }
        },
        "required": [
          "data"
        ],
        "title": "BookingEnvelope",
        "type": "object"
      },
      "BookingListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/BookingOut"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "BookingListEnvelope",
        "type": "object"
      },
      "BookingOut": {
        "properties": {
          "client_name": {
            "title": "Client Name",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "creator": {
            "$ref": "#/components/schemas/BookingCreatorOut"
          },
          "delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "event_city": {
            "title": "Event City",
            "type": "string"
          },
          "event_date": {
            "format": "date",
            "title": "Event Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "quoted_price_idr": {
            "title": "Quoted Price Idr",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "event_date",
          "event_city",
          "notes",
          "quoted_price_idr",
          "created_at",
          "started_at",
          "delivered_at",
          "completed_at",
          "creator",
          "client_name"
        ],
        "title": "BookingOut",
        "type": "object"
      },
...
      "ConversationOut": {
        "properties": {
          "booking_id": {
            "format": "uuid",
            "title": "Booking Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "booking_id",
          "created_at"
        ],
        "title": "ConversationOut",
        "type": "object"
      },
      "CreateBookingRequest": {
        "properties": {
          "creator_id": {
            "format": "uuid",
            "title": "Creator Id",
            "type": "string"
          },
          "event_city": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Event City",
            "type": "string"
          },
          "event_date": {
            "format": "date",
            "title": "Event Date",
            "type": "string"
          },
          "notes": {
            "default": "",
            "maxLength": 2000,
            "title": "Notes",
            "type": "string"
          }
        },
        "required": [
          "creator_id",
          "event_date",
          "event_city"
        ],
        "title": "CreateBookingRequest",
        "type": "object"
      },
...
      "DeliverableOut": {
        "properties": {
          "booking_id": {
            "format": "uuid",
            "title": "Booking Id",
            "type": "string"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "downloadable": {
            "title": "Downloadable",
            "type": "boolean"
          },
          "external_host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Host"
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          },
          "replaces_deliverable_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replaces Deliverable Id"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "source_type": {
            "enum": [
              "private_file",
              "external_link"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "upload_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Id"
          },
          "uploaded_by_user_id": {
            "format": "uuid",
            "title": "Uploaded By User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "booking_id",
          "uploaded_by_user_id",
          "title",
          "description",
          "source_type",
          "upload_id",
          "external_url",
          "external_host",
          "media_type",
          "filename",
          "content_type",
          "size_bytes",
          "replaces_deliverable_id",
          "downloadable",
          "created_at"
        ],
        "title": "DeliverableOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkspaceBookingOut": {
        "additionalProperties": false,
        "properties": {
          "client_name": {
            "title": "Client Name",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "creator": {
            "$ref": "#/components/schemas/BookingCreatorOut"
          },
          "delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "event_city": {
            "title": "Event City",
            "type": "string"
          },
          "event_date": {
            "format": "date",
            "title": "Event Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "quoted_price_idr": {
            "title": "Quoted Price Idr",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "requested",
              "accepted",
              "awaiting_payment",
              "confirmed",
              "in_progress",
              "delivered",
              "rejected",
              "completed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "event_date",
          "event_city",
          "notes",
          "quoted_price_idr",
          "created_at",
          "started_at",
          "delivered_at",
          "completed_at",
          "creator",
          "client_name"
        ],
        "title": "WorkspaceBookingOut",
        "type": "object"
      },
      "WorkspaceEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WorkspaceOut"
          }
        },
        "required": [
          "data"
        ],
        "title": "WorkspaceEnvelope",
        "type": "object"
      },
      "WorkspaceOut": {
        "additionalProperties": false,
        "properties": {
          "booking": {
            "$ref": "#/components/schemas/WorkspaceBookingOut"
          },
          "conversation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConversationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "deliverables": {
            "items": {
              "$ref": "#/components/schemas/DeliverableOut"
            },
            "title": "Deliverables",
            "type": "array"
          },
          "payment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspacePaymentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "enum": [
              "client",
              "creator"
            ],
            "title": "Role",
            "type": "string"
          },
          "unread_count": {
            "minimum": 0.0,
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "role",
          "booking",
          "conversation",
          "deliverables",
          "unread_count",
          "payment"
        ],
        "title": "WorkspaceOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Jepret API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/bookings": {
      "get": {
        "operationId": "list_my_bookings_api_v1_bookings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List My Bookings",
        "tags": [
          "bookings"
        ]
      },
      "post": {
        "operationId": "create_booking_api_v1_bookings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Booking",
        "tags": [
          "bookings"
        ]
      }
    },
...
    "/api/v1/bookings/{booking_id}": {
      "get": {
        "operationId": "get_booking_api_v1_bookings__booking_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Booking",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/accept": {
      "post": {
        "operationId": "accept_booking_api_v1_bookings__booking_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Booking",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/cancel": {
      "post": {
        "operationId": "cancel_booking_api_v1_bookings__booking_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Booking",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/complete": {
      "post": {
        "operationId": "complete_booking_api_v1_bookings__booking_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Booking",
        "tags": [
          "bookings"
        ]
      }
    },
...
    "/api/v1/bookings/{booking_id}/reject": {
      "post": {
        "operationId": "reject_booking_api_v1_bookings__booking_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Booking",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/start": {
      "post": {
        "operationId": "start_booking_api_v1_bookings__booking_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Booking",
        "tags": [
          "bookings"
        ]
      }
    },
...
    "/api/v1/bookings/{booking_id}/workspace": {
      "get": {
        "operationId": "get_booking_workspace_api_v1_bookings__booking_id__workspace_get",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Booking Workspace",
        "tags": [
          "bookings"
        ]
      }
    },
...
  }
}